requirements.txt
annotated-types==0.7.0
anyio==4.5.2
appnope==0.1.4
asttokens==2.4.1
backcall==0.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
flask==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.7.1
jupyter-client==8.6.3
jupyter-core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.55.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
werkzeug==3.0.6
zipp==3.20.2
